requirements.txt
fastapi~=0.120.0
uvicorn[standard]
aiogram~=3.19.0
sqlalchemy~=2.0.40
psycopg-binary
pydantic~=2.10.6
pyyaml~=6.0.2
openpyxl
pandas~=2.3.3
pytest
requests~=2.32.5

pydantic-settings~=2.11.0